% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{add_to_sets}
\alias{add_to_sets}
\title{Add identifier to files from existing sets}
\usage{
add_to_sets(
  set_names,
  new_identifier,
  source_folder = NULL,
  archive_originals = TRUE,
  ...
)
}
\arguments{
\item{set_names}{Names of sets to combine (as shown by list_file_sets)}

\item{new_identifier}{New identifier to add}

\item{source_folder}{Source folder (default: checks both data/source and data/cleaned)}

\item{archive_originals}{Whether to archive the original files after creating new set (default: FALSE)}

\item{...}{Other parameters passed to create_file_set}
}
\description{
Gets all files from specified sets and adds another identifier
}
