% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{advance_to_next_stage}
\alias{advance_to_next_stage}
\title{Advance project to next stage}
\usage{
advance_to_next_stage(project_code, decision = NULL, next_steps = NULL)
}
\arguments{
\item{project_code}{Project code to advance}

\item{decision}{Decision from current stage}

\item{next_steps}{Note about next steps}
}
\description{
Moves a project through the Blitzschreiben pipeline
}
