% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{annotated_bibliography}
\alias{annotated_bibliography}
\title{Create annotated bibliography template in Quarto}
\usage{
annotated_bibliography(include_pdfs = TRUE, only_incomplete = FALSE)
}
\arguments{
\item{include_pdfs}{Link to PDFs if available}

\item{only_incomplete}{Only include entries that need attention}
}
\description{
Create annotated bibliography template in Quarto
}
