% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{apply_workflows}
\alias{apply_workflows}
\title{Run multiple workflows on same file set}
\usage{
apply_workflows(workflow_names, file_set, ...)
}
\arguments{
\item{workflow_names}{Vector of workflow names}

\item{file_set}{File set to process}

\item{...}{Additional arguments passed to apply_workflow}
}
\value{
Named list of results
}
\description{
Run multiple workflows on same file set
}
