% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{archive_files}
\alias{archive_files}
\title{Move files to the old folder}
\usage{
archive_files(filenames, from_folder = NULL, create_date_folder = TRUE)
}
\arguments{
\item{filenames}{Vector of filenames to move}

\item{from_folder}{Source folder path (default: data/source)}

\item{create_date_folder}{Create a subfolder in old/ with today's date (default: TRUE)}
}
\value{
Invisible vector of new file paths
}
\description{
Moves specified files from their current location to the project's old folder
}
