% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{blitzschreiben_status}
\alias{blitzschreiben_status}
\title{Show complete Blitzschreiben pipeline status}
\usage{
blitzschreiben_status(base_path = NULL)
}
\arguments{
\item{base_path}{Base path containing all projects}
}
\description{
Displays all projects across all stages with current status
}
