% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{clean_demographics}
\alias{clean_demographics}
\title{Clean demographic variables}
\usage{
clean_demographics(data, age_var = NULL, gender_var = NULL, race_var = NULL)
}
\arguments{
\item{data}{Data frame containing demographic variables}

\item{age_var}{Name of age variable}

\item{gender_var}{Name of gender variable}

\item{race_var}{Name of race/ethnicity variable}
}
\value{
Data frame with cleaned demographics
}
\description{
Applies standard cleaning to common demographic variables
}
