% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{combine_workflow_results}
\alias{combine_workflow_results}
\title{Combine workflow results based on type}
\usage{
combine_workflow_results(results)
}
\arguments{
\item{results}{List of results from each file}
}
\value{
Combined results
}
\description{
Combine workflow results based on type
}
