% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{create_component_templates}
\alias{create_component_templates}
\title{Create component analysis script templates}
\usage{
create_component_templates(
  project_code,
  analysis_components,
  file_sets,
  project_path
)
}
\arguments{
\item{project_code}{Project code}

\item{analysis_components}{Vector of component names}

\item{file_sets}{File sets detected for the project}

\item{project_path}{Path to project}
}
\value{
Vector of component script paths
}
\description{
Generates individual component scripts called by master script
}
