% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{create_master_script}
\alias{create_master_script}
\title{Create master analysis script for Stage II}
\usage{
create_master_script(
  project_code,
  analysis_components = NULL,
  file_sets = NULL,
  project_path = NULL
)
}
\arguments{
\item{project_code}{Project code}

\item{analysis_components}{Vector of analysis component names}

\item{file_sets}{Named list of file sets to process}

\item{project_path}{Path to project (optional)}
}
\value{
Path to master script
}
\description{
Generates orchestrating script that calls component analysis scripts
}
