% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_meeting_agenda}
\alias{create_meeting_agenda}
\title{Create meeting agenda template}
\usage{
create_meeting_agenda(
  project_code = NULL,
  meeting_type = "progress",
  attendees = NULL
)
}
\arguments{
\item{project_code}{Project code}

\item{meeting_type}{Type of meeting: "kickoff", "progress", "results", "planning"}

\item{attendees}{Vector of attendee names}
}
\value{
Path to agenda template file
}
\description{
Generates structured agenda for project meetings
}
