% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{create_onboarding_navigation}
\alias{create_onboarding_navigation}
\title{Create navigation document for onboarding package}
\usage{
create_onboarding_navigation(
  project_code,
  collaborator_name,
  package_type,
  output_dir
)
}
\arguments{
\item{project_code}{Project code}

\item{collaborator_name}{Name of collaborator}

\item{package_type}{Type of package created}

\item{output_dir}{Directory to save document}
}
\description{
Main entry point that explains what's in the package
}
