% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{detect_id_columns}
\alias{detect_id_columns}
\title{Detect likely ID columns in dataset}
\usage{
detect_id_columns(data)
}
\arguments{
\item{data}{Data frame to analyze}
}
\value{
Vector of likely ID column names
}
\description{
Identifies columns that appear to be identifier variables
}
