% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{detect_required_packages}
\alias{detect_required_packages}
\title{Detect required packages from workflow code}
\usage{
detect_required_packages(code_lines)
}
\arguments{
\item{code_lines}{Lines of workflow code}
}
\value{
Character vector of required packages
}
\description{
Detect required packages from workflow code
}
