% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{determine_current_stage}
\alias{determine_current_stage}
\title{Determine current project stage}
\usage{
determine_current_stage(project_code, paths)
}
\arguments{
\item{project_code}{Project code}

\item{paths}{Project paths}
}
\value{
Stage description
}
\description{
Figures out what stage the project is currently in
}
