% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup hook}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Automatically sets up exit logging when PRoMAn is loaded
}
