% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{extract_keyword_from_set_name}
\alias{extract_keyword_from_set_name}
\title{Extract keyword from file set name for loading}
\usage{
extract_keyword_from_set_name(set_name)
}
\arguments{
\item{set_name}{Name of the file set}
}
\value{
Keyword string for file loading
}
\description{
Helper function to determine what keyword to use for file loading
}
