% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communication.R
\name{extract_stage1_info}
\alias{extract_stage1_info}
\title{Extract Stage I information from project}
\usage{
extract_stage1_info(project_code, paths)
}
\arguments{
\item{project_code}{Project code}

\item{paths}{Project paths}
}
\value{
List of extracted information
}
\description{
Pulls information from Stage I triage card if it exists
}
