% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallback.R
\name{fallback_info}
\alias{fallback_info}
\title{Show information about current fallback}
\usage{
fallback_info(path = NULL, show_files = FALSE)
}
\arguments{
\item{path}{Project directory (default: NULL uses current project)}

\item{show_files}{Logical, whether to list all backed up files (default: FALSE)}
}
\description{
Show information about current fallback
}
