% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{file_exists}
\alias{file_exists}
\title{Check if file exists in project}
\usage{
file_exists(filename, folder = NULL)
}
\arguments{
\item{filename}{Filename to check}

\item{folder}{Folder to check in (default: checks all)}
}
\value{
TRUE if file exists, FALSE otherwise
}
\description{
Check if file exists in project
}
