% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{find_stale_files}
\alias{find_stale_files}
\title{Find files that haven't been modified recently}
\usage{
find_stale_files(days = 30, folder_path = NULL)
}
\arguments{
\item{days}{Number of days to consider "stale" (default: 30)}

\item{folder_path}{Folder to check (default: data/source)}
}
\value{
Vector of stale filenames
}
\description{
Identifies stale files that might need cleanup
}
