% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_management.R
\name{format_academic_table}
\alias{format_academic_table}
\title{Format data frame for academic table presentation}
\usage{
format_academic_table(data, caption = "", note = "", round_digits = 2)
}
\arguments{
\item{data}{Data frame to format}

\item{caption}{Table caption}

\item{note}{Table note or footer}

\item{round_digits}{Number of decimal places for rounding}
}
\value{
Formatted data frame
}
\description{
Applies academic formatting standards to tables
}
