% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{generate_data_dictionary}
\alias{generate_data_dictionary}
\title{Generate comprehensive data dictionary}
\usage{
generate_data_dictionary(
  data,
  filename = NULL,
  include_stats = TRUE,
  guess_descriptions = FALSE
)
}
\arguments{
\item{data}{Data frame to document}

\item{filename}{Optional filename to save dictionary}

\item{include_stats}{Include descriptive statistics}

\item{guess_descriptions}{Attempt to guess variable descriptions}
}
\value{
Data dictionary data frame
}
\description{
Creates documentation for all variables in dataset
}
