% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{get_folder_summary}
\alias{get_folder_summary}
\title{Get summary of folder contents}
\usage{
get_folder_summary(folder_path = NULL, sort_by = "date")
}
\arguments{
\item{folder_path}{Path to folder (default: data/source)}

\item{sort_by}{Sort by: "name", "size", or "date" (default: "date")}
}
\value{
Data frame with file information
}
\description{
Returns information about files in a folder including size and modification dates
}
