% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_logging.R
\name{infer_recent_activity}
\alias{infer_recent_activity}
\title{Infer what activities have been happening based on file changes}
\usage{
infer_recent_activity(days_back = 7, paths = NULL)
}
\arguments{
\item{days_back}{Number of days to analyze}

\item{paths}{Project paths (optional, will get if not provided)}
}
\value{
Character vector with activity inferences
}
\description{
Analyzes recent changes to guess what type of work has been done
}
