% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_setup.R
\name{list_files}
\alias{list_files}
\title{List files in project folder with minimal syntax}
\usage{
list_files(folder = "source", pattern = NULL, full_names = FALSE)
}
\arguments{
\item{folder}{Folder name}

\item{pattern}{Optional pattern to match}

\item{full_names}{Return full paths (default: FALSE)}
}
\value{
Vector of filenames
}
\description{
List files in project folder with minimal syntax
}
