% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{list_workflows}
\alias{list_workflows}
\title{List available workflows}
\usage{
list_workflows(show_details = TRUE)
}
\arguments{
\item{show_details}{Show detailed information}
}
\value{
Data frame of workflows (invisibly)
}
\description{
Shows all registered workflows with their descriptions
}
