% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{parse_bib_entries}
\alias{parse_bib_entries}
\title{Parse BibTeX entries helper function}
\usage{
parse_bib_entries(bib_content)
}
\arguments{
\item{bib_content}{Lines from .bib file}
}
\description{
Enhanced parser that includes start/end positions
}
