% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{parse_filename_metadata}
\alias{parse_filename_metadata}
\title{Parse metadata from filename}
\usage{
parse_filename_metadata(filename)
}
\arguments{
\item{filename}{Filename without extension}
}
\description{
Attempts to extract author, year, title from common filename patterns
}
