% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{process_files_parallel}
\alias{process_files_parallel}
\title{Process files in parallel}
\usage{
process_files_parallel(
  workflow,
  files,
  save_outputs,
  workflow_name,
  n_cores = NULL,
  verbose = TRUE
)
}
\arguments{
\item{workflow}{Workflow object}

\item{files}{Vector of file paths}

\item{save_outputs}{Whether to save outputs}

\item{workflow_name}{Name of workflow}

\item{n_cores}{Number of cores (NULL for auto-detect)}

\item{verbose}{Show progress}
}
\value{
List of results
}
\description{
Process files in parallel
}
