% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{read_stage1_card}
\alias{read_stage1_card}
\title{Read Stage I card information}
\usage{
read_stage1_card(project_code)
}
\arguments{
\item{project_code}{Project code}
}
\value{
List with card information
}
\description{
Extracts key information from Stage I triage card in personal_notes
}
