% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_management.R
\name{register_figure}
\alias{register_figure}
\title{Register figure in metadata registry}
\usage{
register_figure(
  figure_name,
  filename_base,
  figure_type,
  caption,
  notes,
  formats,
  width,
  height
)
}
\arguments{
\item{figure_name}{Original figure name}

\item{filename_base}{Base filename used}

\item{figure_type}{Type of figure}

\item{caption}{Caption text}

\item{notes}{Notes about figure}

\item{formats}{Formats saved}

\item{width}{Figure width}

\item{height}{Figure height}
}
\description{
Internal function to track figure metadata
}
