% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{register_workflow_internal}
\alias{register_workflow_internal}
\title{Register workflow for reuse}
\usage{
register_workflow_internal(workflow)
}
\arguments{
\item{workflow}{Workflow object from extraction}
}
\description{
Register workflow for reuse
}
