% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{save_combined_outputs}
\alias{save_combined_outputs}
\title{Save combined outputs}
\usage{
save_combined_outputs(combined, output_name, workflow_name)
}
\arguments{
\item{combined}{Combined results}

\item{output_name}{Base name for outputs}

\item{workflow_name}{Workflow name}
}
\description{
Save combined outputs
}
