% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{setup_naming_helpers}
\alias{setup_naming_helpers}
\title{Set up naming helpers in environment}
\usage{
setup_naming_helpers(env, pattern, workflow_name)
}
\arguments{
\item{env}{Environment to modify}

\item{pattern}{Naming pattern}

\item{workflow_name}{Workflow name}
}
\description{
Set up naming helpers in environment
}
