% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{show_workflow}
\alias{show_workflow}
\title{Show workflow details including I/O analysis}
\usage{
show_workflow(workflow_name)
}
\arguments{
\item{workflow_name}{Name of workflow to examine}
}
\description{
Show workflow details including I/O analysis
}
