% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{stage3_progress_check}
\alias{stage3_progress_check}
\title{Stage III progress and word count tracker}
\usage{
stage3_progress_check(project_code)
}
\arguments{
\item{project_code}{Project code}
}
\description{
Shows completion status and word counts for draft
}
