% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{tag_pdfs}
\alias{tag_pdfs}
\title{Tag PDFs for organization}
\usage{
tag_pdfs(pdf_files = "all", tags)
}
\arguments{
\item{pdf_files}{Character vector of PDF filenames or "all"}

\item{tags}{Character vector of tags (e.g., "methods", "background")}
}
\description{
Tag PDFs for organization
}
