% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{test_workflow}
\alias{test_workflow}
\title{Quick workflow test on single file}
\usage{
test_workflow(workflow_name, test_file, verbose = TRUE)
}
\arguments{
\item{workflow_name}{Workflow to test}

\item{test_file}{Single file to test on}

\item{verbose}{Show detailed output}
}
\value{
Result from workflow
}
\description{
Test a workflow on one file before applying to entire set
}
