% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blitzschreiben.R
\name{track_sources}
\alias{track_sources}
\title{Track source collection progress}
\usage{
track_sources(search_strategy = NULL, new_sources = 0, project_code = NULL)
}
\arguments{
\item{search_strategy}{Description of search approach}

\item{new_sources}{Number of new sources found}

\item{project_code}{Project code (optional)}
}
\description{
Monitors PDF sources in standard sources folder
}
