% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{translate_bibliography}
\alias{translate_bibliography}
\title{Export bibliography to other formats}
\usage{
translate_bibliography(format = c("ris", "csv", "endnote"))
}
\arguments{
\item{format}{Output format: "ris", "csv", or "endnote"}
}
\description{
Export bibliography to other formats
}
