% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_management.R
\name{update_bibliography}
\alias{update_bibliography}
\title{Update bibliography by merging all sources and incorporating annotations}
\usage{
update_bibliography()
}
\description{
\enumerate{
\item Checks for annotation file and incorporates if found
\item Merges any other .bib files in sources folder
\item Removes duplicates
\item Updates the main bibliography file
}
}
