% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{validate_file_set}
\alias{validate_file_set}
\title{Validate file set}
\usage{
validate_file_set(files)
}
\arguments{
\item{files}{Vector of file paths}
}
\value{
Validated file paths
}
\description{
Validate file set
}
