% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{workflow_exists}
\alias{workflow_exists}
\title{Check if a workflow exists}
\usage{
workflow_exists(workflow_name)
}
\arguments{
\item{workflow_name}{Name of workflow to check}
}
\value{
TRUE if workflow exists
}
\description{
Check if a workflow exists
}
